orchestrator/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
celery[redis]==5.4.0
redis==5.2.1
pydantic==2.10.4
pydantic-settings==2.7.1
rebound==4.4.3
reboundx==4.2.1
qutip==5.0.4
numpy==1.26.4
websockets==14.1
pyscf==2.7.0
MDAnalysis==2.8.0
pybullet==3.2.6
einsteinpy==0.4.0
copasi-basico==0.75
tellurium==2.2.10
brian2==2.7.1
msprime==1.3.3
rdkit-pypi==2024.3.5
cantera==3.0.1
sympy==1.13.3
gmsh==4.13.1
lcapy==1.23
pennylane==0.39.0
matplotlib==3.9.3
control==0.10.1
pyomo==6.8.2
networkx==3.4.2
phiflow==2.7.0
highspy==1.7.2
cyipopt==1.4.1
jax[cpu]==0.4.35
vtk==9.3.1
tskit>=0.5.8
pyslim>=1.0.4
simuPOP>=1.1.17
reportlab>=4.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
pytest>=8.0
